CONST_versions_requirements.txt
Babel==2.3.4
boto==2.43.0
bottle==0.12.13
c2c.template==1.3.0
Chameleon==2.25
dogpile.cache==0.6.2
fanstatic==0.16
GeoAlchemy2==0.4.0
GeoFormAlchemy2==2.0.dev3
geojson==1.3.3
httplib2==0.9.2
ipcalc==1.99.0
isodate==0.5.4
Jinja2==2.8
lingua==4.11
Mako==1.0.6
Markdown==2.6.7
MarkupSafe==0.23
OWSLib==0.13.0
papyrus==2.2
Paste==2.0.3
PasteDeploy==1.5.2
PasteScript==2.0.2
Pillow==3.4.2
polib==1.0.8
postmarkup==1.2.2
psycopg2==2.7.5
pykwalify==1.5.2
pyproj==1.9.5.1
pyramid-chameleon==0.3
pyramid-debugtoolbar==3.0.5
pyramid-mako==1.0.2
pyramid-multiauth==0.9.0
pyramid-tm==1.1.1
pyramid_closure==1.0.0
python-dateutil==2.6.0
pytz==2016.7
repoze.lru==0.6
requests==2.12.1
simplejson==3.10.0
six==1.10.0
SQLAHelper==1.0
SQLAlchemy==1.1.4
Tempita==0.5.2
textile==2.3.6
tilecloud==0.3.0
tilecloud-chain==1.2.1
transaction==2.0.3
translationstring==1.3
venusian==1.0
waitress==1.0.1
WebError==0.13.1
WebHelpers==1.3
WebOb==1.6.3
zope.component==4.3.0
zope.deprecation==4.2.0
zope.event==4.2.0
zope.interface==4.3.2
zope.sqlalchemy==0.7.7
